+++ FlashCap.Core/Internal/NativeMethods_AVFoundation.cs
@@ -7,7 +7,7 @@ using System.Threading;
 
 namespace FlashCap.Internal;
 
-internal static class NativeMethods_AVFoundation
+internal static partial class NativeMethods_AVFoundation
 {
     public static readonly Dictionary<PixelFormats, int> PixelFormatMap = new()
     {
@@ -853,706 +853,7 @@ internal static class NativeMethods_AVFoundation
             s[3];*/
     }
 
-    public static class LibAVFoundation
-    {
-        public const string Path = "/System/Library/Frameworks/AVFoundation.framework/AVFoundation";
-
-        public static readonly IntPtr Handle = Dlfcn.OpenLibrary(Path, Dlfcn.Mode.Now);
-
-        public delegate void AVRequestAccessStatus(bool accessGranted);
-
-        public sealed class AVFrameRateRange : LibObjC.NSObject
-        {
-            public AVFrameRateRange(IntPtr handle, bool retain) :
-                base(handle, retain)
-            { }
-
-            public LibCoreMedia.CMTime MaxFrameDuration
-            {
-                get
-                {
-                    if (LibSystem.IsOnArm64)
-                    {
-                        return LibObjC.SendAndGetCMTime(Handle, LibObjC.GetSelector("maxFrameDuration"));
-                    }
-
-                    LibObjC.SendAndGetCMTimeStret(out var result, Handle, LibObjC.GetSelector("maxFrameDuration"));
-
-                    return result;
-                }
-            }
-
-            public LibCoreMedia.CMTime MinFrameDuration
-            {
-                get
-                {
-                    if (LibSystem.IsOnArm64)
-                    {
-
-                        return LibObjC.SendAndGetCMTime(Handle, LibObjC.GetSelector("minFrameDuration"));
-
-                    }
-
-
-                    LibObjC.SendAndGetCMTimeStret(out var result, Handle, LibObjC.GetSelector("minFrameDuration"));
-
-                    return result;
-                }
-            }
-
-        }
-
-        public sealed class AVCaptureDevice : LibObjC.NSObject
-        {
-            public AVCaptureDevice(IntPtr handle, bool retain) :
-                base(handle, retain)
-            { }
-
-            public string UniqueID
-            {
-                get
-                {
-                    using var result = new LibCoreFoundation.CFString(
-                        LibObjC.SendAndGetHandle(
-                            Handle,
-                            LibObjC.GetSelector("uniqueID")));
-
-                    var str = result.ToString() ?? throw new InvalidOperationException();
-
-                    return str;
-                }
-            }
-
-            public string ModelID
-            {
-                get
-                {
-                    using var result = new LibCoreFoundation.CFString(
-                        LibObjC.SendAndGetHandle(
-                            Handle,
-                            LibObjC.GetSelector("modelID")));
-
-                    return result.ToString() ?? throw new InvalidOperationException();
-                }
-            }
-
-            public string LocalizedName
-            {
-                get
-                {
-                    using var result = new LibCoreFoundation.CFString(
-                        LibObjC.SendAndGetHandle(
-                            Handle,
-                            LibObjC.GetSelector("localizedName")));
-
-                    return result.ToString() ?? throw new InvalidOperationException();
-                }
-
-            }
-
-            public AVCaptureDeviceFormat[] Formats
-            {
-                get
-                {
-                    var handle = LibObjC.SendAndGetHandle(
-                        Handle,
-                        LibObjC.GetSelector("formats"));
-
-                    return LibCoreFoundation.CFArray.ToArray(handle, static handle => new AVCaptureDeviceFormat(handle, retain: true));
-                }
-            }
-
-            public AVCaptureDeviceFormat ActiveFormat
-            {
-                get => new AVCaptureDeviceFormat(
-                    LibObjC.SendAndGetHandle(
-                        Handle,
-                        LibObjC.GetSelector("activeFormat")),
-                    retain: true);
-                set =>
-                    LibObjC.SendNoResult(
-                        Handle,
-                        LibObjC.GetSelector("setActiveFormat:"),
-                        value.Handle);
-            }
-
-            public LibCoreMedia.CMTime ActiveVideoMinFrameDuration
-            {
-                get
-                {
-                    if (LibSystem.IsOnArm64)
-                    {
-                        LibObjC.SendAndGetCMTimeStret(out var result, Handle, LibObjC.GetSelector("activeVideoMinFrameDuration"));
-
-                        return result;
-                    }
-
-                    return LibObjC.SendAndGetCMTime(Handle, LibObjC.GetSelector("activeVideoMinFrameDuration"));
-                }
-                set
-                {
-                    LibObjC.SendNoResult(
-                        Handle,
-                        LibObjC.GetSelector("setActiveVideoMinFrameDuration:"),
-                        value);
-                }
-            }
-
-            public LibCoreMedia.CMTime ActiveVideoMaxFrameDuration
-            {
-                get
-                {
-                    if (LibSystem.IsOnArm64)
-                    {
-                        LibObjC.SendAndGetCMTimeStret(out var result, Handle, LibObjC.GetSelector("activeVideoMaxFrameDuration"));
-
-                        return result;
-                    }
-
-                    return LibObjC.SendAndGetCMTime(Handle, LibObjC.GetSelector("activeVideoMaxFrameDuration"));
-                }
-                set
-                {
-                    LibObjC.SendNoResult(
-                        Handle,
-                        LibObjC.GetSelector("setActiveVideoMaxFrameDuration:"),
-                        value);
-                }
-            }
-
-            public unsafe void LockForConfiguration()
-            {
-                IntPtr errorHandle;
-                LibObjC.SendNoResult(
-                    Handle,
-                    LibObjC.GetSelector("lockForConfiguration:"),
-                    new IntPtr(&errorHandle));
-
-                if (errorHandle != IntPtr.Zero)
-                {
-                    using var error = new LibObjC.NSError(errorHandle, retain: true);
-                    Console.WriteLine(error.Error);
-                    throw new InvalidOperationException(/* error.FailureReason */);
-                }
-            }
-
-            public unsafe void UnlockForConfiguration()
-            {
-                LibObjC.SendNoResult(
-                    Handle,
-                    LibObjC.GetSelector("unlockForConfiguration"));
-            }
-
-            public static AVCaptureDevice? DeviceWithUniqueID(string deviceUniqueID)
-            {
-                using var nativeDeviceUniqueID = LibCoreFoundation.CFString.Create(deviceUniqueID);
-                var handle = LibObjC.SendAndGetHandle(
-                    LibObjC.GetClass(nameof(AVCaptureDevice)),
-                    LibObjC.GetSelector("deviceWithUniqueID:"),
-                    nativeDeviceUniqueID.Handle);
-
-                return handle == IntPtr.Zero ? null : new AVCaptureDevice(handle, retain: false);
-            }
-
-            public static AVAuthorizationStatus GetAuthorizationStatus(IntPtr mediaType)
-            {
-                return (AVAuthorizationStatus)(long)
-                    LibObjC.SendAndGetHandle(
-                        LibObjC.GetClass(nameof(AVCaptureDevice)),
-                        LibObjC.GetSelector("authorizationStatusForMediaType:"),
-                        mediaType);
-            }
-
-            private static LibObjC.BlockLiteralFactory? RequestAccessForMediaTypeBlockFactory;
-            private delegate void RequestAccessForMediaTypeTrampoline(IntPtr block, byte accessGranted);
-
-            public static unsafe void RequestAccessForMediaType(IntPtr mediaType, AVRequestAccessStatus completion)
-            {
-                RequestAccessForMediaTypeBlockFactory ??= LibObjC.BlockLiteralFactory.CreateFactory<RequestAccessForMediaTypeTrampoline>(
-                    signature: "v@?^vC",
-                    delegate (IntPtr block, byte accessGranted)
-                    {
-                        LibObjC.BlockLiteral
-                            .GetTarget<AVRequestAccessStatus>(block)
-                            .Invoke(accessGranted != 0);
-                    });
-
-                using var blockLiteral = RequestAccessForMediaTypeBlockFactory.CreateLiteral(completion);
-
-                LibObjC.SendNoResult(
-                    LibObjC.GetClass(nameof(AVCaptureDevice)),
-                    LibObjC.GetSelector("requestAccessForMediaType:completionHandler:"),
-                    mediaType,
-                    new IntPtr(&blockLiteral));
-            }
-        }
-
-        public sealed class AVCaptureDeviceDiscoverySession : LibObjC.NSObject
-        {
-            public AVCaptureDeviceDiscoverySession(IntPtr handle, bool retain) :
-                base(handle, retain)
-            { }
-
-            public AVCaptureDevice[] Devices =>
-                LibCoreFoundation.CFArray.ToArray(
-                    LibObjC.SendAndGetHandle(
-                        Handle,
-                        LibObjC.GetSelector("devices")),
-                    static handle => new AVCaptureDevice(handle, retain: true));
-
-            public static AVCaptureDeviceDiscoverySession DiscoverySessionWithVideoDevices()
-            {
-                var deviceTypes = new[]
-                {
-                    AVCaptureDeviceType.BuiltInWideAngleCamera,
-                    AVCaptureDeviceType.BuiltInTelephotoCamera,
-                    AVCaptureDeviceType.BuiltInDualCamera,
-                    AVCaptureDeviceType.BuiltInTripleCamera,
-                    AVCaptureDeviceType.BuiltInTrueDepthCamera,
-                    AVCaptureDeviceType.ExternalCamera,
-                    AVCaptureDeviceType.ExternalContinuityCamera
-                }
-                    .Where(static handle => handle != IntPtr.Zero)
-                    .ToArray();
-
-                using var deviceTypesArray = LibCoreFoundation.CFArray.Create(deviceTypes);
-
-                var mediaType = AVMediaType.Video;
-                var position = (long)AVCaptureDevicePosition.Unspecified;
-
-                return new AVCaptureDeviceDiscoverySession(
-                    LibObjC.SendAndGetHandle(
-                        LibObjC.GetClass(nameof(AVCaptureDeviceDiscoverySession)),
-                        LibObjC.GetSelector("discoverySessionWithDeviceTypes:mediaType:position:"),
-                        deviceTypesArray.Handle,
-                        mediaType,
-                        position),
-                    retain: true);
-            }
-        }
-
-        public sealed class AVCaptureDeviceFormat : LibObjC.NSObject
-        {
-            public AVCaptureDeviceFormat(IntPtr handle, bool retain) :
-                base(handle, retain)
-            { }
-
-            public LibCoreMedia.CMFormatDescription FormatDescription =>
-                new LibCoreMedia.CMFormatDescription(
-                    LibObjC.SendAndGetHandle(
-                        Handle,
-                        LibObjC.GetSelector("formatDescription")));
-
-            public AVFrameRateRange[] VideoSupportedFrameRateRanges =>
-                LibCoreFoundation.CFArray.ToArray(
-                    LibObjC.SendAndGetHandle(
-                        Handle,
-                        LibObjC.GetSelector("videoSupportedFrameRateRanges")),
-                    static handle => new AVFrameRateRange(handle, retain: true));
-        }
-
-        public enum AVCaptureDevicePosition : long
-        {
-            Unspecified = 0,
-            Back = 1,
-            Front = 2,
-        }
-
-        public static class AVCaptureDeviceType
-        {
-            public static readonly IntPtr BuiltInWideAngleCamera = Dlfcn.GetSymbolIndirect(LibAVFoundation.Handle, "AVCaptureDeviceTypeBuiltInWideAngleCamera");
-            public static readonly IntPtr BuiltInTelephotoCamera = Dlfcn.GetSymbolIndirect(LibAVFoundation.Handle, "AVCaptureDeviceTypeBuiltInTelephotoCamera");
-            public static readonly IntPtr BuiltInDualCamera = Dlfcn.GetSymbolIndirect(LibAVFoundation.Handle, "AVCaptureDeviceTypeBuiltInDualCamera");
-            public static readonly IntPtr BuiltInTripleCamera = Dlfcn.GetSymbolIndirect(LibAVFoundation.Handle, "AVCaptureDeviceTypeBuiltInTripleCamera");
-            public static readonly IntPtr BuiltInTrueDepthCamera = Dlfcn.GetSymbolIndirect(LibAVFoundation.Handle, "AVCaptureDeviceTypeBuiltInTrueDepthCamera");
-            public static readonly IntPtr ExternalCamera = Dlfcn.GetSymbolIndirect(LibAVFoundation.Handle, "AVCaptureDeviceTypeExternal");
-            public static readonly IntPtr ExternalContinuityCamera = Dlfcn.GetSymbolIndirect(LibAVFoundation.Handle, "AVCaptureDeviceTypeContinuityCamera");
-        }
-
-        public abstract class AVCaptureInput : LibObjC.NSObject
-        {
-            protected AVCaptureInput(IntPtr handle, bool retain) :
-                base(handle, retain)
-            { }
-        }
-
-        public abstract class AVCaptureOutput : LibObjC.NSObject
-        {
-            protected AVCaptureOutput(IntPtr handle, bool retain) :
-                base(handle, retain)
-            { }
-        }
-
-        public sealed class AVCaptureDeviceInput : AVCaptureInput
-        {
-            public AVCaptureDeviceInput(AVCaptureDevice device) :
-                base(FromDevice(device), retain: true)
-            { }
-
-            private static unsafe IntPtr FromDevice(AVCaptureDevice device)
-            {
-                var errorHandle = default(IntPtr);
-                var inputHandle = LibObjC.SendAndGetHandle(
-                    LibObjC.GetClass("AVCaptureDeviceInput"),
-                    LibObjC.GetSelector("deviceInputWithDevice:error:"),
-                    device.Handle,
-                    new IntPtr(&errorHandle));
-
-                if (errorHandle != IntPtr.Zero)
-                {
-                    using var error = new LibObjC.NSError(errorHandle, retain: false);
-                    Console.WriteLine(error.Error);
-                    throw new InvalidOperationException(/* error.FailureReason */);
-                }
-
-                return inputHandle;
-            }
-        }
-
-        public sealed class AVCaptureVideoDataOutput : AVCaptureOutput
-        {
-            public AVCaptureVideoDataOutput() : base(IntPtr.Zero, retain: false)
-            {
-                Init();
-            }
-
-            private void Init()
-            {
-
-                var videoAlloc = LibObjC.SendAndGetHandle(
-                    LibObjC.GetClass("AVCaptureVideoDataOutput"),
-                    LibObjC.GetSelector(LibObjC.AllocSelector));
-                
-                
-                Handle = LibObjC.SendAndGetHandle(
-                    videoAlloc,
-                    LibObjC.GetSelector("init"));
-            }
-
-            public unsafe int[] AvailableVideoCVPixelFormatTypes =>
-                LibCoreFoundation.CFArray.ToArray(
-                    LibObjC.SendAndGetHandle(
-                        Handle,
-                        LibObjC.GetSelector("availableVideoCVPixelFormatTypes")),
-                    static handle =>
-                    {
-                        int value;
-                        if (LibCoreFoundation.CFNumberGetValue(handle, LibCoreFoundation.CFNumberType.sInt32Type, &value))
-                            return value;
-                        throw new InvalidOperationException("The value contained by CFNumber cannot be read as 32-bit signed integer.");
-                    });
-
-            public bool AlwaysDiscardsLateVideoFrames
-            {
-                get =>
-                    LibObjC.SendAndGetBool(
-                        Handle,
-                        LibObjC.GetSelector("alwaysDiscardsLateVideoFrames"));
-                set =>
-                    LibObjC.SendNoResult(
-                        Handle,
-                        LibObjC.GetSelector("setAlwaysDiscardsLateVideoFrames:"),
-                        value);
-            }
-
-            public unsafe void SetPixelFormatType(int format)
-            {
-                format = 1111970369;
-                
-                //var number = LibCoreFoundation.CFNumberCreate(IntPtr.Zero, LibCoreFoundation.CFNumberType.sInt32Type, &format);
-                
-                var nsNumber = LibObjC.CreateNSNumber(format);
-                
-                LibCoreFoundation.CFRetain(nsNumber);
-                
-                //var keys = new IntPtr[] { LibCoreVideo.kCVPixelBufferPixelFormatTypeKey};
-                //var values = new IntPtr[] { nsNumber };
-
-                /*var dictionary = LibCoreFoundation.CFDictionaryCreate(
-                    IntPtr.Zero,
-                    keys,
-                    values,
-                    numValues: 1,
-                    LibCoreFoundation.kCFCopyStringDictionaryKeyCallBacks,
-                    LibCoreFoundation.kCFTypeDictionaryValueCallBacks);*/
-
-                IntPtr nsDictionaryClass = LibCoreFoundation.BasicClassesHelper.NSDictionary;
-                IntPtr dictSel =  LibObjC.GetSelector("dictionaryWithObject:forKey:");
-                
-               //LibCoreFoundation.CFRetain(dictionary);
-
-                IntPtr nsPixelFormatKey = LibCoreVideo.kCVPixelBufferPixelFormatTypeKey;
-               
-                IntPtr videoSettings =  LibObjC.SendAndGetHandle(nsDictionaryClass, dictSel, nsNumber, nsPixelFormatKey);
-                LibCoreFoundation.CFRetain(videoSettings);
-                IntPtr setVideoSettingsSel = LibObjC.GetSelector("setVideoSettings:");
-                
-                LibObjC.SendNoResult(
-                    Handle,
-                    setVideoSettingsSel,
-                    videoSettings);
-
-                LibCoreFoundation.CFRelease(videoSettings);
-                LibCoreFoundation.CFRelease(nsNumber);
-                //LibCoreFoundation.CFRelease(strTrue.Handle);
-            }
-
-            public void SetSampleBufferDelegate(AVCaptureVideoDataOutputSampleBuffer sampleBufferDelegate, LibCoreFoundation.DispatchQueue sampleBufferCallbackQueue) =>
-                LibObjC.SendNoResult(
-                    Handle,
-                    LibObjC.GetSelector("setSampleBufferDelegate:queue:"),
-                    sampleBufferDelegate.Handle,
-                    sampleBufferCallbackQueue.Handle);
-        }
-
-        public abstract class AVCaptureVideoDataOutputSampleBuffer : LibObjC.NSObject
-        {
-            private const string HandleVariableName = nameof(GCHandle);
-            private static IntPtr HandleVariableDescriptor;
-
-            static AVCaptureVideoDataOutputSampleBuffer()
-            {
-                var handle = LibObjC.AllocateClass(
-                    LibObjC.GetClass("NSObject"),
-                    nameof(AVCaptureVideoDataOutputSampleBuffer),
-                    extraBytes: IntPtr.Zero);
-
-                LibObjC.AddMethod(
-                    handle,
-                    LibObjC.GetSelector("dealloc"),
-                    Marshal.GetFunctionPointerForDelegate(DeallocTrampoline),
-                    types: "v@:");
-
-                LibObjC.AddMethod(
-                    handle,
-                    LibObjC.GetSelector("captureOutput:didDropSampleBuffer:fromConnection:"),
-                    Marshal.GetFunctionPointerForDelegate(DidDropSampleBufferTrampoline),
-                    types: "v@:@@@");
-
-                LibObjC.AddMethod(
-                    handle,
-                    LibObjC.GetSelector("captureOutput:didOutputSampleBuffer:fromConnection:"),
-                    Marshal.GetFunctionPointerForDelegate(DidOutputSampleBufferTrampoline),
-                    types: "v@:@@@");
-
-                LibObjC.AddProtocol(
-                    handle,
-                    LibObjC.GetProtocol("AVCaptureVideoDataOutputSampleBufferDelegate"));
-
-                LibObjC.AddVariable(
-                    handle,
-                    HandleVariableName,
-                    size: new IntPtr(IntPtr.Size),
-                    alignment: IntPtr.Size switch
-                    {
-                        4 => 2,
-                        8 => 3,
-                        _ => throw new NotSupportedException("The current arhitecture isn't supported.")
-                    },
-                    types: "^v");
-
-                LibObjC.RegisterClass(handle);
-
-                HandleVariableDescriptor = LibObjC.GetVariable(handle, HandleVariableName);
-            }
-
-            protected AVCaptureVideoDataOutputSampleBuffer() : base(
-                LibObjC.SendAndGetHandle(
-                    LibObjC.GetClass(nameof(AVCaptureVideoDataOutputSampleBuffer)),
-                    LibObjC.GetSelector(LibObjC.AllocSelector)),
-                retain: false)
-            {
-                LibObjC.SetVariable(
-                    Handle,
-                    HandleVariableDescriptor,
-                    (IntPtr)GCHandle.Alloc(this));
-            }
-
-            public abstract void DidDropSampleBuffer(IntPtr captureOutput, IntPtr sampleBuffer, IntPtr connection);
-
-            public abstract void DidOutputSampleBuffer(IntPtr captureOutput, IntPtr sampleBuffer, IntPtr connection);
-
-            private delegate void DidDropSampleBufferDelegate(IntPtr self, IntPtr captureOutput, IntPtr sampleBuffer, IntPtr connection);
-            public delegate void DidOutputSampleBufferDelegate(IntPtr self, IntPtr captureOutput, IntPtr sampleBuffer, IntPtr connection);
-            private delegate void DeallocDelegate(IntPtr self);
-
-            private static DidDropSampleBufferDelegate DidDropSampleBufferTrampoline = (self, captureOutput, sampleBuffer, connection) =>
-            {
-                var handle = LibObjC.GetVariable(self, HandleVariableDescriptor);
-                var obj = GCHandle.FromIntPtr(handle).Target as AVCaptureVideoDataOutputSampleBuffer;
-
-                obj?.DidDropSampleBuffer(captureOutput, sampleBuffer, connection);
-            };
-
-            public static DidOutputSampleBufferDelegate DidOutputSampleBufferTrampoline = (self, captureOutput, sampleBuffer, connection) =>
-            {
-                var handle = LibObjC.GetVariable(self, HandleVariableDescriptor);
-                var obj = GCHandle.FromIntPtr(handle).Target as AVCaptureVideoDataOutputSampleBuffer;
-
-                obj?.DidOutputSampleBuffer(captureOutput, sampleBuffer, connection);
-            };
-
-            private static DeallocDelegate DeallocTrampoline = (self) =>
-            {
-                var handle = LibObjC.GetVariable(self, HandleVariableDescriptor);
-
-                GCHandle
-                    .FromIntPtr(handle)
-                    .Free();
-            };
-        }
-
-        public sealed class AVCaptureSession : LibObjC.NSObject
-        {
-            public AVCaptureSession() : base(IntPtr.Zero, false)
-            {
-                Init();
-            }
-
-            private void Init()
-            {
-                var sessionClass = LibObjC.SendAndGetHandle(
-                    LibObjC.GetClass("AVCaptureSession"),
-                    LibObjC.GetSelector(LibObjC.AllocSelector));
-                
-                var sessionObj = LibObjC.SendAndGetHandle(
-                    sessionClass,
-                    LibObjC.GetSelector("init"));
-                
-                Handle = sessionObj;
-                
-                LibCoreFoundation.CFRetain(this.Handle);
-                
-            }
-
-            public void AddInput(AVCaptureInput input) =>
-                LibObjC.SendNoResult(
-                    Handle,
-                    LibObjC.GetSelector("addInput:"),
-                    input.Handle);
-
-            public void AddOutput(AVCaptureOutput output)
-            {
-                IntPtr allocSel = LibObjC.GetSelector("alloc");
-                IntPtr initSel = LibObjC.GetSelector("init");
-            
-                // Criação da saída de vídeo: AVCaptureVideoDataOutput.
-                IntPtr videoDataOutputClass = LibObjC.GetClass("AVCaptureVideoDataOutput");
-                IntPtr videoOutputAlloc = LibObjC.SendAndGetHandle(videoDataOutputClass, allocSel);
-                IntPtr videoDataOutput = LibObjC.SendAndGetHandle(videoOutputAlloc, initSel);    
-                
-                IntPtr pixelFormatTypeKeyPtr =  Dlfcn.dlsym(LibCoreVideo.Handle, "kCVPixelBufferPixelFormatTypeKey");
-                if (pixelFormatTypeKeyPtr == IntPtr.Zero)
-                {
-                    Console.WriteLine("Não foi possível obter kCVPixelBufferPixelFormatTypeKey.");
-                    return;
-                }
-                
-                // Obtém o valor real da NSString a partir da constante
-                IntPtr nsPixelFormatKey = Marshal.ReadIntPtr(pixelFormatTypeKeyPtr);
-                
-                int pixelFormat = 1111970369; // Exemplo utilizando BGRA
-                IntPtr nsNumber = LibObjC.CreateNSNumber(pixelFormat);
-                
-                // Utiliza a chave reconhecida pelo Core Video: kCVPixelBufferPixelFormatTypeKey
-                //IntPtr nsPixelFormatKey = CreateNSString("kCVPixelBufferPixelFormatTypeKey");
-                
-                IntPtr nsDictionaryClass = LibObjC.GetClass("NSDictionary");
-                IntPtr dictSel = LibObjC.GetSelector("dictionaryWithObject:forKey:");
-                IntPtr videoSettings = LibObjC.SendAndGetHandle(nsDictionaryClass, dictSel, nsNumber, nsPixelFormatKey);
-                IntPtr setVideoSettingsSel = LibObjC.GetSelector("setVideoSettings:");
-                LibObjC.SendNoResult(videoDataOutput, setVideoSettingsSel, videoSettings);
-
-                // Criação e registro da classe delegate dinâmica que implementa o protocolo AVCaptureVideoDataOutputSampleBufferDelegate.
-                IntPtr nsObjectClass = LibObjC.GetClass("NSObject");
-                IntPtr delegateClass = LibObjC.objc_allocateClassPair(nsObjectClass, "CaptureDelegate", IntPtr.Zero);
-                // Seleciona o método a ser implementado.
-                IntPtr selDidOutput = LibObjC.GetSelector("captureOutput:didOutputSampleBuffer:fromConnection:");
-                
-                
-                
-                CaptureOutputDidOutputSampleBuffer callbackDelegate = new CaptureOutputDidOutputSampleBuffer(CaptureOutputCallback);
-                
-                
-                IntPtr impCallback = Marshal.GetFunctionPointerForDelegate(callbackDelegate);
-            
-                // A string de tipos "v@:@@@" indica um método que retorna void e recebe (self, _cmd, output, sampleBuffer, connection).
-                string types = "v@:@@@";
-                bool added = LibObjC.class_addMethod(delegateClass, selDidOutput, impCallback, types);
-                if (!added)
-                {
-                    Console.WriteLine("Falha ao adicionar método ao delegate.");
-                    return;
-                }
-                LibObjC.objc_registerClassPair(delegateClass);
-
-                // Cria uma instância do delegate.
-                IntPtr delegateInstanceAlloc = LibObjC.SendAndGetHandle(delegateClass, allocSel);
-                IntPtr delegateInstance = LibObjC.SendAndGetHandle(delegateInstanceAlloc, initSel);
-
-                // Cria uma fila de despacho (dispatch queue) para os callbacks.
-                IntPtr dispatchQueue = LibObjC.dispatch_queue_create("VideoMovie", IntPtr.Zero);
-
-                // Define o delegate para a saída de vídeo:
-                // [videoDataOutput setSampleBufferDelegate:delegateInstance queue:dispatchQueue]
-                IntPtr setDelegateSel = LibObjC.GetSelector("setSampleBufferDelegate:queue:");
-                LibObjC.SendNoResult(videoDataOutput, setDelegateSel, delegateInstance, dispatchQueue);
-                
-                LibObjC.SendNoResult(
-                    Handle,
-                    LibObjC.GetSelector("addOutput:"),
-                    videoDataOutput);
-                
-                
-            }
-
-            // Delegate correspondente ao método:
-            // - (void)captureOutput:(id)output didOutputSampleBuffer:(id)sampleBuffer fromConnection:(id)connection;
-            [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-            public delegate void CaptureOutputDidOutputSampleBuffer(IntPtr self, IntPtr _cmd, IntPtr output, IntPtr sampleBuffer, IntPtr connection);
-
-            
-            public static void CaptureOutputCallback(IntPtr self, IntPtr _cmd, IntPtr output, IntPtr sampleBuffer, IntPtr connection)
-            {
-                Console.WriteLine("Frame capturado.");
-                // Processamento adicional do sampleBuffer pode ser implementado aqui.
-            
-                var pixelBuffer = LibCoreMedia.CMSampleBufferGetImageBuffer(sampleBuffer);
-            
-                Console.WriteLine("PixelBuffer: " + pixelBuffer);
-            }
-            
-            public bool CanAddOutput(AVCaptureOutput output) =>
-                LibObjC.SendAndGetBool(
-                    Handle,
-                    LibObjC.GetSelector("canAddOutput:"),
-                    output.Handle);
-
-            public void StartRunning() =>
-                LibObjC.SendNoResult(
-                    Handle,
-                    LibObjC.GetSelector("startRunning"));
-
-            public void StopRunning() =>
-                LibObjC.SendNoResult(
-                    Handle,
-                    LibObjC.GetSelector("stopRunning"));
-        }
-
-        public enum AVAuthorizationStatus : long
-        {
-            NotDetermined,
-            Restricted,
-            Denied,
-            Authorized
-        }
-
-        public static class AVMediaType
-        {
-            public static readonly IntPtr Video = Dlfcn.GetSymbolIndirect(LibAVFoundation.Handle, "AVMediaTypeVideo");
-            //public static readonly IntPtr Video = Dlfcn.GetSymbolIndirect(LibAVFoundation.Handle, "AVMediaTypeVideo");
-        }
-    }
+    
 
     internal abstract class NativeObject : IDisposable
     {
